Should be done with my Sprint 0.5. Specifically, the following has been completed:
1) Cleaned up the 2D Maze Generation code. Should be quite modular, and in
particular easily modified to allow special rooms and different room
sizes. Only hangups are in later interfacing with the class.

2) Wrote fresh code to draw the maze in 3D. Currently has a few fiddly
components - in particular, corridors are REALLY long because the grid
size has been scaled up a lot to allow room for block placement - but
this should be easilly enough addressed later. Also not the cleanest code,
but I got pressed for time, so I guess I'll just have to clean it up
later.

3) Put in proper quad and cube models from Blender. Unfortunately, I'm not
the most expert at using these softwares, so textures look really terrible
for the time being. I'll look into it when I have a spare moment, but I
doubt it will be working for the walking skeleton demo.