Rooms have specified types now, so you can decide what to put in a room based on the type. Currently has empty, enemy, and start room types.
Enemies getting placed in enemy typed rooms. Roughly speaking happens on
1/2 of the rooms. No spawning in corridors.

Player now spawns in the center of whichever room is specified as the
start room. Technically there's a bug here: if we, through sheer bad luck,
have no rooms spawn, the user will just start wherever (potentially out of
the maze). Proper fix for this is ensuring there will always be some
number of rooms, but that's fine tuning in the maze generation we can
address later.