Fixed up block placement, so they are now along the walls of rooms instead of random piles in the floor. Specifically, they show up in proper rooms, rather than also in the middle of small corridor intersections as well. I also added a second ore type, just to give a bit more visual appeal. We should be able to at least hack together a solution to make ores mine for resources and stone blocks to mine for nothing by Friday no problem, I figure.
Also, I finally figured out how to properly assign materials to objects
that have been imported, so I've fiddled around to make things look
somewhat nicer. Doesn't evoke the most amazing graphics, but substantially
better than flat colors we had before.

Still no lights, because I still haven't figured out how to properly spawn
lights.

Finally, fiddled around with the map scaling somewhat. Rooms are
substantially smaller now, and the ceiling is lower down. Decided to go
with that, because the blocks were taking a VERY long time to generate
without it. Long term, we'll probably want to up the size again, but
before that we'll want to fix up the game so we only load some parts of
the maze at a time, speeding up that process. But when my fancy desktop
machine is taking a minute or two to generate all the cubes, there is
definitely going to be problems on demo day without scaling back.