Cube placement in rooms seems to be working, sans openings in walls for corridors.
Don't recall if this happened in this update or the previous, but for the
sake of documentation: scaling is no longer happening (or, rather it is,
but it's currently scaling by 1). I had stuck to the premise that I needed
to scale to "unity coordinates", when really my coordinates were based on
something in Unity to begin with: the mineable blocks. By getting rid of
scaling, all my weird placement issues went away. Well, that and I needed
to transpose some coordinates but that was not a big deal.

Anyway, next iteration is to put in gaps in the block placement to allow
passage to corridors, as well as cleaning up any code relating to scaling
so the scaling is no longer there.