Figured out why the floor and ceiling were blacked out. I misinterpreted the meaning of "main color" as the main color of the object. Apparently, this is actually the main color of the -light- as assigned to the object. So, if you want your texture to show up, you need it to be white. Whoops. However, still won't capture light from point light. Also figured out this issue (or at least, I'm fairly confident I did), but I haven't been able to actually fix it, because I'm having issues importing a corrected model. I'll keep fiddling to see if I can get it to work, but just being able to see the ceiling and floor on OUYA will be a major improvement.
As a heads up: I haven't actually had a chance to test this on the device
yet. I'm confident this should fix the issue (since I started encountering
the issue on my laptop as I fiddled with materials, and it fixed it
there), but I'll double check at home tonight.